Add sleep to openai and refactor
moved evaluator to controller and cleaned up code
